Edited gitignore for swap, renamed tsr.py and __init__
