+++ .gitignore
@@ -7,3 +7,4 @@
 **/cmovetraj.txt
 
 doc
+*~
